Implemented Adding of Components in Details Panel + Bunch of other stuff
Already exists warning message with component type and game object name.
Re-ordered contents of GO AddComponent to work with physics bodies being added when a collider isn't present on the object before hand (Used to break and throw up error logs)
Reworked FindFirstActiveCamera function to get the first active camera as opposed to only using a predetermined play camera, adds a  default object with a camera if all cameras in scene are disabled and we are in play mode.