Integrate ImGuizmo library into FluxEditor build
Added ImGuizmo as a dependency in the FluxEditor project, updated CMake and Premake scripts to build and link ImGuizmo, and included the necessary library files for both debug and release configurations. Also improved CMake configuration for SFML and ImGui-SFML discovery. This enables usage of ImGuizmo in the editor and streamlines its build process.